knossos.history/complete: use return values always
We used to use the invocation value for invocations preferentially, only
replacing them when they were nil. This is not great when values are
data structures with unknown components at invocation time, e.g. a
multi-key read. Now we always take the completion value, and
additionally enforce that failure have the same invocation and failure
values.